Worked over the autoplay restriction of mobile chrome.
The mobile browsers do not allow the the auto play of the audio/video elements.
You can't even try to trigger play() programatically until and unless it is
preceeded by a user click. The solution here is to show the play button when
song is loaded enough to start playing and then the user can click on the play
button. After this multiple songs can be played programatically.